requirements.txt
astroid==2.3.3
bcrypt==3.1.7
cffi==1.14.0
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.0.0
pkg-resources==0.0.0
pycparser==2.19
pylint==2.4.4
six==1.14.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
Werkzeug==0.16.1
wrapt==1.11.2
WTForms==2.2.1
astroid==2.3.3
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.0.0
pkg-resources==0.0.0
pycparser==2.19
pylint==2.4.4
six==1.14.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
Werkzeug==0.16.1
wrapt==1.11.2
WTForms==2.2.1
